+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* [improvement] More verbose output for redeploy, start and stop
+
 ## 0.2.12 / 2013-03-25
 
 * [bug] shelly check recognizes puma as allowed web server